+++ R/download.R
@@ -1,9 +1,11 @@
-#' @importFrom httr2 request
-#' @importFrom httr2 req_progress
-#' @importFrom httr2 req_perform
 mydownload <- function(url, destfile) {
-    req <- request(url) |> 
-        req_progress()
-    
-    req |> req_perform(path = destfile)
+    if (is.installed('httr2')) {    
+        req <- httr2::request(url) |> 
+            httr2::req_progress()
+        
+        req |> httr2::req_perform(path = destfile)
+    } else {
+        download.file(url = url, destfile = destfile)
+    }
 }
+